requirements.txt
# ── Embeddings (Gemini) ────────────────────────────────────────
google-generativeai==0.8.4
 
# ── LLM (Groq — free, no CC) ───────────────────────────────────
groq==0.23.1
 
# ── Vector store (Pinecone hosted, free tier) ──────────────────
pinecone==5.4.2

# ── Chunking ───────────────────────────────────────────────────
chonkie[semantic]==1.6.0

# ── Crawling & parsing ─────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.13.3
trafilatura==2.0.0

# ── Eval ───────────────────────────────────────────────────────
ragas==0.2.15
datasets==3.5.0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.1.0
tiktoken==0.9.0
tenacity==9.1.2
rich==14.0.0
tqdm==4.67.1

pytest==8.3.5
pytest-asyncio==0.25.3     # for testing any async functions in pipeline.py